restructure browsergui.elements into basic/input/layout submodules
