+++ browsergui/elements/basic/__init__.py
@@ -1,0 +1,9 @@
+'''Defines simple, static, atomic GUI elements:
+
+- :class:`.Text` for plain text, and a few subclasses (:class:`.Link`, :class:`.CodeSnippet`, :class:`.Paragraph`, :class:`.CodeBlock`)
+- :class:`.Image`
+
+'''
+
+from ._text import *
+from ._image import Image
