+++ browsergui/elements/input/__init__.py
@@ -1,0 +1,20 @@
+'''Defines GUI elements that gather input from the user:
+
+- :class:`.Button`
+- :class:`.Slider` (abstract) for draggable sliders, and a few subclasses (:class:`.FloatSlider` for floats, :class:`.IntegerSlider` for integers)
+- :class:`.TextField` (single-line) and :class:`.BigTextField` (multi-line)
+- :class:`.NumberField`
+- :class:`.Dropdown`
+- :class:`.ColorField`
+- :class:`.DateField`
+'''
+
+from ._button import Button
+from ._inputfield import InputField
+from ._textfield import TextField
+from ._bigtextfield import BigTextField
+from ._dropdown import Dropdown
+from ._numberfield import NumberField
+from ._colorfield import ColorField
+from ._datefield import DateField
+from ._slider import Slider, FloatSlider, IntegerSlider
