+++ browsergui/elements/input/_textfield.py
@@ -1,5 +1,5 @@
-from .input_field import InputField
-from ..pythoncompatibility import STRING_TYPES
+from ._inputfield import InputField
+from ...pythoncompatibility import STRING_TYPES
 
 class TextField(InputField):
   '''A single-line text input field.
