+++ browsergui/elements/layout/__init__.py
@@ -1,0 +1,12 @@
+'''Defines GUI elements that arrange their children in certain ways:
+
+- :class:`.Container`, a very simple element with no fancy layout stuff, meant to group other elements together (e.g. to put multiple Paragraph elements as a single List item)
+- :class:`.List`, a bulleted/numbered list of elements
+- :class:`.Grid`
+- :class:`.Viewport`, a small scrollable window viewing a large element
+'''
+
+from ._container import Container
+from ._grid import Grid
+from ._list import List
+from ._viewport import Viewport
