+++ browsergui/elements/layout/_container.py
@@ -1,5 +1,5 @@
-from . import Element
-from ..pythoncompatibility import collections_abc
+from .. import Element
+from ...pythoncompatibility import collections_abc
 
 class Container(Element, collections_abc.MutableSequence):
   """Contains other elements without any fancy layout stuff.
