+++ doc/browsergui.elements.basic.rst
@@ -1,0 +1,10 @@
+browsergui.elements.basic package
+=================================
+
+Module contents
+---------------
+
+.. automodule:: browsergui.elements.basic
+    :members:
+    :undoc-members:
+    :show-inheritance:
