+++ doc/browsergui.elements.input.rst
@@ -1,0 +1,10 @@
+browsergui.elements.input package
+=================================
+
+Module contents
+---------------
+
+.. automodule:: browsergui.elements.input
+    :members:
+    :undoc-members:
+    :show-inheritance:
