+++ doc/browsergui.elements.layout.rst
@@ -1,0 +1,10 @@
+browsergui.elements.layout package
+==================================
+
+Module contents
+---------------
+
+.. automodule:: browsergui.elements.layout
+    :members:
+    :undoc-members:
+    :show-inheritance:
