+++ doc/browsergui.elements.rst
@@ -1,137 +1,14 @@
 browsergui.elements package
 ===========================
 
-Submodules
-----------
+Subpackages
+-----------
 
-browsergui.elements.bigtextfield module
----------------------------------------
+.. toctree::
 
-.. automodule:: browsergui.elements.bigtextfield
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.button module
----------------------------------
-
-.. automodule:: browsergui.elements.button
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.color_field module
---------------------------------------
-
-.. automodule:: browsergui.elements.color_field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.container module
-------------------------------------
-
-.. automodule:: browsergui.elements.container
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.date_field module
--------------------------------------
-
-.. automodule:: browsergui.elements.date_field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.dropdown module
------------------------------------
-
-.. automodule:: browsergui.elements.dropdown
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.grid module
--------------------------------
-
-.. automodule:: browsergui.elements.grid
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.image module
---------------------------------
-
-.. automodule:: browsergui.elements.image
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.input_field module
---------------------------------------
-
-.. automodule:: browsergui.elements.input_field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.link module
--------------------------------
-
-.. automodule:: browsergui.elements.link
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.list module
--------------------------------
-
-.. automodule:: browsergui.elements.list
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.number_field module
----------------------------------------
-
-.. automodule:: browsergui.elements.number_field
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.slider module
----------------------------------
-
-.. automodule:: browsergui.elements.slider
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.text module
--------------------------------
-
-.. automodule:: browsergui.elements.text
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.textfield module
-------------------------------------
-
-.. automodule:: browsergui.elements.textfield
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-browsergui.elements.viewport module
------------------------------------
-
-.. automodule:: browsergui.elements.viewport
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
+    browsergui.elements.basic
+    browsergui.elements.input
+    browsergui.elements.layout
 
 Module contents
 ---------------
